refactor: useOnAuthStateChanged + Router component
